-预入库单导入修改行号类型; -出库单导入新增数量的换算插入;

+++ wms-dao/src/main/java/com/wms/mybatis/dao/DocAsnDetailsMybatisDao.java
@@ -16,7 +16,7 @@ import java.util.Map;
  */
 public interface DocAsnDetailsMybatisDao extends BaseDao {
 	
-	long getAsnlinenoById(DocAsnDetailQuery docAsnDetailQuery);
+	int getAsnlinenoById(DocAsnDetailQuery docAsnDetailQuery);
 	
 	void receiveByAsn(DocAsnDetail docAsnDetail);
